Binomial and Poisson Confidence Intervals
Added the ability calculate probability of individual success given a
binomial distribution, a number of successes, a sample size, and a
cumulative probability.  The result was another function allowing
confidence interval calculation of an estimated p for the binomial
distribution.

Additionally, a function allowing the calculation of the average number
of successes given a known number of successes and a cumulative
probability was added.  The result being a confidence interval for the
point estimation of the average number of successes assuming a poisson
distribution.